Fixed a bug while replaying edits and addressed review comments.
