+++ hadoop-hdfs-project/hadoop-hdfs/src/main/java/org/apache/hadoop/hdfs/server/namenode/FSDirSnapshotOp.java
@@ -17,22 +17,17 @@
  */
 package org.apache.hadoop.hdfs.server.namenode;
 
-import com.google.common.collect.Lists;
 import org.apache.hadoop.HadoopIllegalArgumentException;
 import org.apache.hadoop.fs.InvalidPathException;
 import org.apache.hadoop.fs.Path;
-import org.apache.hadoop.fs.XAttr;
-import org.apache.hadoop.fs.XAttrSetFlag;
 import org.apache.hadoop.fs.permission.FsAction;
 import org.apache.hadoop.hdfs.DFSUtil;
-import org.apache.hadoop.hdfs.XAttrHelper;
 import org.apache.hadoop.hdfs.protocol.FSLimitException;
 import org.apache.hadoop.hdfs.protocol.HdfsFileStatus;
 import org.apache.hadoop.hdfs.protocol.SnapshotDiffReport;
 import org.apache.hadoop.hdfs.protocol.SnapshotDiffReportListing;
 import org.apache.hadoop.hdfs.protocol.SnapshotException;
 import org.apache.hadoop.hdfs.protocol.SnapshottableDirectoryStatus;
-import org.apache.hadoop.hdfs.server.common.HdfsServerConstants;
 import org.apache.hadoop.hdfs.server.namenode.FSDirectory.DirOp;
 import org.apache.hadoop.hdfs.server.namenode.snapshot.DirectorySnapshottableFeature;
 import org.apache.hadoop.hdfs.server.namenode.snapshot.Snapshot;
@@ -45,7 +40,6 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
-import java.util.EnumSet;
 
 class FSDirSnapshotOp {
   /** Verify if the snapshot name is legal. */
@@ -257,38 +251,6 @@ class FSDirSnapshotOp {
 
     // time of snapshot deletion
     final long now = Time.now();
-    if (fsd.isSnapshotDeletionOrdered()) {
-      final INodeDirectory srcRoot = snapshotManager.getSnapshottableRoot(iip);
-      final DirectorySnapshottableFeature snapshottable
-          = srcRoot.getDirectorySnapshottableFeature();
-      final Snapshot snapshot = snapshottable.getSnapshotByName(
-          srcRoot, snapshotName);
-
-      // Diffs must be not empty since a snapshot exists in the list
-      final int earliest = snapshottable.getDiffs().iterator().next()
-          .getSnapshotId();
-      if (snapshot.getId() != earliest) {
-        final XAttr snapshotXAttr = buildXAttr();
-        final List<XAttr> xattrs = Lists.newArrayListWithCapacity(1);
-        xattrs.add(snapshotXAttr);
-
-        // The snapshot to be deleted is just marked for deletion in the xAttr.
-        // Same snaphot delete call can happen multiple times until annd unless
-        // the very 1st instance of a snapshot delete hides it/remove it from
-        // snapshot list. XAttrSetFlag.REPLACE needs to be set to here in order
-        // to address this.
-
-        // Xattr will set on the snapshot root directory
-        FSDirXAttrOp.unprotectedSetXAttrs(fsd,
-            INodesInPath.append(iip, snapshot.getRoot(),
-                DFSUtil.string2Bytes(snapshotName)), xattrs,
-            EnumSet.of(XAttrSetFlag.CREATE, XAttrSetFlag.REPLACE));
-        fsd.getEditLog().logDeleteSnapshot(snapshotRoot, snapshotName,
-            logRetryCache, now);
-        return null;
-      }
-    }
-
     final INode.BlocksMapUpdateInfo collectedBlocks = deleteSnapshot(
         fsd, snapshotManager, iip, snapshotName, now);
     fsd.getEditLog().logDeleteSnapshot(snapshotRoot, snapshotName,
@@ -378,9 +340,5 @@ class FSDirSnapshotOp {
     if (sm.getNumSnapshottableDirs() > 0) {
       checkSnapshot(iip.getLastINode(), snapshottableDirs);
     }
-  }
-
-  public static XAttr buildXAttr() {
-    return XAttrHelper.buildXAttr(HdfsServerConstants.SNAPSHOT_XATTR_NAME);
   }
 }
